requirements.txt
asgiref==3.3.4
Django==3.2
pytz==2021.1
sqlparse==0.4.1
channels==3.0.3
channels_redis==3.2.0
djangorestframework==3.12.4
